+++ doc/crypto-driver/appendix/history.rst
@@ -1,4 +1,5 @@
 .. SPDX-FileCopyrightText: Copyright 2025 Arm Limited and/or its affiliates
+.. SPDX-FileCopyrightText: Copyright 2026 GlobalPlatform
 .. SPDX-License-Identifier: CC-BY-SA-4.0 AND LicenseRef-Patent-license
 
 Changes to the API
@@ -17,6 +18,6 @@ Changes in the draft GlobalPlatform publication revision
 Other changes
 ~~~~~~~~~~~~~
 
-*   Migrated the document to the 2026 PSA Certified API template.
+*   Migrated the document to the 2026 PSA API template.
 
     This changes the document front matter structure and publication styling, without changing the API.